requirements.txt
numpy>=1.20
sympy
scipy
matplotlib
networkx
requests
pyYAML
ordered-set
scikits.ode